+++ test/v2/addLiquidityNested/addLiquidityNested.integration.test.ts
@@ -11,40 +11,22 @@ import {
     publicActions,
     PublicActions,
     TestActions,
-    TransactionReceipt,
     WalletActions,
     walletActions,
 } from 'viem';
 
+import { NestedPoolState } from '@/entities';
+import { Address, Hex } from '@/types';
+import { CHAINS, ChainId } from '@/utils';
+
+import { ANVIL_NETWORKS, startFork } from 'test/anvil/anvil-global-setup';
+import { POOLS, TestToken, TOKENS } from 'test/lib/utils/addresses';
 import {
-    Slippage,
-    AddLiquidityNested,
-    replaceWrapped,
-    TokenAmount,
-    NestedPoolState,
-} from '../src/entities';
-import { Address, Hex } from '../src/types';
-
-import { BALANCER_RELAYER, CHAINS, ChainId } from '../src/utils';
-
-import { forkSetup, sendTransactionGetBalances } from './lib/utils/helper';
-import { Relayer } from '../src/entities/relayer';
-import { AddLiquidityNestedInput } from '../src/entities/addLiquidityNested/types';
-import { ANVIL_NETWORKS, startFork } from './anvil/anvil-global-setup';
-import { POOLS, TestToken, TOKENS } from './lib/utils/addresses';
-
-type TxInput = {
-    poolId: Hex;
-    amountsIn: {
-        address: Address; // DAI
-        rawAmount: bigint;
-    }[];
-    chainId: ChainId;
-    rpcUrl: string;
-    testAddress: Address;
-    client: Client & PublicActions & TestActions & WalletActions;
-    useNativeAssetAsWrappedAmountIn?: boolean;
-};
+    assertResults,
+    doAddLiquidityNested,
+} from 'test/lib/utils/addLiquidityNestedHelper';
+import { forkSetup } from 'test/lib/utils/helper';
+import { AddLiquidityNestedTxInput } from 'test/lib/utils/types';
 
 const chainId = ChainId.MAINNET;
 const DAI = TOKENS[chainId].DAI;
@@ -61,6 +43,7 @@ describe('add liquidity nested test', () => {
     let testAddress: Address;
     let mainTokens: TestToken[];
     let initialBalances: bigint[];
+    let txInput: AddLiquidityNestedTxInput;
 
     beforeAll(async () => {
         // setup chain and test client
@@ -78,8 +61,22 @@ describe('add liquidity nested test', () => {
 
         poolId = BPT_WETH_3POOL.id;
 
+        // setup mock api
+        const api = new MockApi();
+        // get pool state from api
+        const nestedPoolState = await api.getNestedPool(poolId);
+
         mainTokens = [WETH, DAI, USDC, USDT];
         initialBalances = mainTokens.map((t) => parseUnits('1000', t.decimals));
+
+        txInput = {
+            nestedPoolState,
+            chainId,
+            rpcUrl,
+            testAddress,
+            client,
+            amountsIn: [],
+        };
     });
 
     beforeEach(async () => {
@@ -101,6 +98,11 @@ describe('add liquidity nested test', () => {
             },
         ];
 
+        txInput = {
+            ...txInput,
+            amountsIn,
+        };
+
         const {
             transactionReceipt,
             balanceDeltas,
@@ -108,14 +110,7 @@ describe('add liquidity nested test', () => {
             minBptOut,
             slippage,
             value,
-        } = await doTransaction({
-            poolId,
-            amountsIn,
-            chainId,
-            rpcUrl,
-            testAddress,
-            client,
-        });
+        } = await doAddLiquidityNested(txInput);
 
         assertResults(
             transactionReceipt,
@@ -124,6 +119,7 @@ describe('add liquidity nested test', () => {
             balanceDeltas,
             slippage,
             minBptOut,
+            chainId,
             value,
         );
     });
@@ -134,6 +130,11 @@ describe('add liquidity nested test', () => {
             rawAmount: parseUnits('1', t.decimals),
         }));
 
+        txInput = {
+            ...txInput,
+            amountsIn,
+        };
+
         const {
             transactionReceipt,
             balanceDeltas,
@@ -141,14 +142,7 @@ describe('add liquidity nested test', () => {
             minBptOut,
             slippage,
             value,
-        } = await doTransaction({
-            poolId,
-            amountsIn,
-            chainId,
-            rpcUrl,
-            testAddress,
-            client,
-        });
+        } = await doAddLiquidityNested(txInput);
 
         assertResults(
             transactionReceipt,
@@ -157,6 +151,7 @@ describe('add liquidity nested test', () => {
             balanceDeltas,
             slippage,
             minBptOut,
+            chainId,
             value,
         );
     });
@@ -169,6 +164,12 @@ describe('add liquidity nested test', () => {
 
         const useNativeAssetAsWrappedAmountIn = true;
 
+        txInput = {
+            ...txInput,
+            amountsIn,
+            useNativeAssetAsWrappedAmountIn,
+        };
+
         const {
             transactionReceipt,
             balanceDeltas,
@@ -176,15 +177,7 @@ describe('add liquidity nested test', () => {
             minBptOut,
             slippage,
             value,
-        } = await doTransaction({
-            poolId,
-            amountsIn,
-            chainId,
-            rpcUrl,
-            testAddress,
-            client,
-            useNativeAssetAsWrappedAmountIn,
-        });
+        } = await doAddLiquidityNested(txInput);
 
         assertResults(
             transactionReceipt,
@@ -193,6 +186,7 @@ describe('add liquidity nested test', () => {
             balanceDeltas,
             slippage,
             minBptOut,
+            chainId,
             value,
             useNativeAssetAsWrappedAmountIn,
         );
@@ -208,127 +202,17 @@ describe('add liquidity nested test', () => {
 
         const useNativeAssetAsWrappedAmountIn = true;
 
-        await expect(() =>
-            doTransaction({
-                poolId,
-                amountsIn,
-                chainId,
-                rpcUrl,
-                testAddress,
-                client,
-                useNativeAssetAsWrappedAmountIn,
-            }),
-        ).rejects.toThrowError(
+        txInput = {
+            ...txInput,
+            amountsIn,
+            useNativeAssetAsWrappedAmountIn,
+        };
+
+        await expect(() => doAddLiquidityNested(txInput)).rejects.toThrowError(
             'Adding liquidity with native asset requires wrapped native asset to exist within amountsIn',
         );
     });
 });
-
-export const doTransaction = async ({
-    poolId,
-    amountsIn,
-    chainId,
-    rpcUrl,
-    testAddress,
-    client,
-    useNativeAssetAsWrappedAmountIn = false,
-}: TxInput) => {
-    // setup mock api
-    const api = new MockApi();
-    // get pool state from api
-    const nestedPoolFromApi = await api.getNestedPool(poolId);
-    // setup add liquidity helper
-    const addLiquidityNested = new AddLiquidityNested();
-
-    const addLiquidityInput: AddLiquidityNestedInput = {
-        amountsIn,
-        chainId,
-        rpcUrl,
-        accountAddress: testAddress,
-        useNativeAssetAsWrappedAmountIn,
-    };
-    const queryOutput = await addLiquidityNested.query(
-        addLiquidityInput,
-        nestedPoolFromApi,
-    );
-
-    // build add liquidity call with expected minBpOut based on slippage
-    const slippage = Slippage.fromPercentage('1'); // 1%
-
-    const signature = await Relayer.signRelayerApproval(
-        BALANCER_RELAYER[chainId],
-        testAddress,
-        client,
-    );
-
-    const { call, to, value, minBptOut } = addLiquidityNested.buildCall({
-        ...queryOutput,
-        slippage,
-        sender: testAddress,
-        recipient: testAddress,
-        relayerApprovalSignature: signature,
-    });
-
-    let tokensIn = queryOutput.amountsIn.map((a) => a.token);
-    if (useNativeAssetAsWrappedAmountIn) {
-        tokensIn = replaceWrapped(tokensIn, chainId);
-    }
-
-    // send add liquidity transaction and check balance changes
-    const { transactionReceipt, balanceDeltas } =
-        await sendTransactionGetBalances(
-            [
-                ...tokensIn.map((t) => t.address),
-                queryOutput.bptOut.token.address,
-            ],
-            client,
-            testAddress,
-            to,
-            call,
-            value,
-        );
-    return {
-        transactionReceipt,
-        balanceDeltas,
-        bptOut: queryOutput.bptOut,
-        minBptOut,
-        slippage,
-        value,
-    };
-};
-
-const assertResults = (
-    transactionReceipt: TransactionReceipt,
-    bptOut: TokenAmount,
-    amountsIn: {
-        address: Address;
-        rawAmount: bigint;
-    }[],
-    balanceDeltas: bigint[],
-    slippage: Slippage,
-    minBptOut: bigint,
-    value?: bigint,
-    useNativeAssetAsWrappedAmountIn = false,
-) => {
-    expect(transactionReceipt.status).to.eq('success');
-    expect(bptOut.amount > 0n).to.be.true;
-    const expectedDeltas = [
-        ...amountsIn.map((a) => a.rawAmount),
-        bptOut.amount,
-    ];
-    expect(expectedDeltas).to.deep.eq(balanceDeltas);
-    const expectedMinBpt = slippage.applyTo(bptOut.amount, -1);
-    expect(expectedMinBpt).to.deep.eq(minBptOut);
-
-    const weth = amountsIn.find(
-        (a) => a.address === '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
-    );
-    if (weth && useNativeAssetAsWrappedAmountIn) {
-        expect(value).to.eq(weth.rawAmount);
-    } else {
-        expect(value).to.eq(undefined || 0n);
-    }
-};
 
 /*********************** Mock To Represent API Requirements **********************/
